+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/text/StringBooleanOp.java
@@ -1,13 +1,15 @@
 package org.enso.table.data.column.operation.map.text;
 
-import java.util.BitSet;
 import org.enso.table.data.column.operation.map.MapOperation;
 import org.enso.table.data.column.storage.BoolStorage;
+import org.enso.table.data.column.storage.SpecializedStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.StringStorage;
 import org.enso.table.error.UnexpectedTypeException;
 
-public abstract class StringBooleanOp extends MapOperation<StringStorage> {
+import java.util.BitSet;
+
+public abstract class StringBooleanOp extends MapOperation<SpecializedStorage<String>> {
   public StringBooleanOp(String name) {
     super(name);
   }
@@ -19,14 +21,13 @@ public abstract class StringBooleanOp extends MapOperation<StringStorage> {
   }
 
   @Override
-  public Storage runMap(StringStorage storage, Object arg) {
+  public Storage runMap(SpecializedStorage<String> storage, Object arg) {
     if (arg == null) {
       BitSet newVals = new BitSet();
       BitSet newMissing = new BitSet();
       newMissing.set(0, storage.size());
       return new BoolStorage(newVals, newMissing, storage.size(), false);
-    } else if (arg instanceof String) {
-      String argString = (String) arg;
+    } else if (arg instanceof String argString) {
       BitSet newVals = new BitSet();
       BitSet newMissing = new BitSet();
       for (int i = 0; i < storage.size(); i++) {
@@ -52,9 +53,8 @@ public abstract class StringBooleanOp extends MapOperation<StringStorage> {
   }
 
   @Override
-  public Storage runZip(StringStorage storage, Storage arg) {
-    if (arg instanceof StringStorage) {
-      StringStorage v = (StringStorage) arg;
+  public Storage runZip(SpecializedStorage<String> storage, Storage arg) {
+    if (arg instanceof StringStorage v) {
       BitSet newVals = new BitSet();
       BitSet newMissing = new BitSet();
       for (int i = 0; i < storage.size(); i++) {