+++ CHANGELOG.md
@@ -188,6 +188,7 @@
 - [Removed `Array.set_at`.][3634]
 - [Added various date part functions to `Date` and `Date_Time`.][3669]
 - [Implemented `Table.take` and `Table.drop` for the in-memory backend.][3647]
+- [Implemented specialized storage for the in-memory Table.][3673]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -299,6 +300,7 @@
 [3634]: https://github.com/enso-org/enso/pull/3634
 [3669]: https://github.com/enso-org/enso/pull/3669
 [3647]: https://github.com/enso-org/enso/pull/3647
+[3673]: https://github.com/enso-org/enso/pull/3673
 
 #### Enso Compiler